+++ src/ducks/index.js
@@ -1,14 +1,10 @@
 import { combineReducers } from 'redux';
-// import { firebaseReducer } from 'react-redux-firebase';
-// import { firestoreReducer } from 'redux-firestore';
+
+// my ducks
 import alertSnackbarReducer from './alertSnackbar';
 import userReducer from './user';
 
-// my ducks
-
 const rootReducer = combineReducers({
-  // firebase: firebaseReducer,
-  // firestore: firestoreReducer,
   user: userReducer,
   alertSnackbar: alertSnackbarReducer,
 });