+++ src/keys/index.js
@@ -1,0 +1,8 @@
+//  API
+export const API_URL = process.env.REACT_APP_API_URL;
+
+// ENCRYPT - DECRYPT
+export const SALT = process.env.REACT_APP_CRYPTO_SALT;
+export const IV = process.env.REACT_APP_CRYPTO_IV;
+export const KEY_PASSWORD = process.env.REACT_APP_CRYPTO_KEY_PASSWORD;
+export const KEY_JSON = process.env.REACT_APP_CRYPTO_KEY_JSON;