+++ src/tools/alertSnackbar.js
@@ -9,6 +9,11 @@ export const USER_SUCCESSFULLY_LOADED = {
   message: 'El usuario fue cargado correctamente',
 };
 
+export const PASSWORD_SENT_TO_EMAIL_SUCCESSFULLY = {
+  severity: 'success',
+  message: 'La contraseña fue enviada a su correo correctamente',
+};
+
 // errors
 export const ENTER_VALID_UNI_CODE = {
   severity: 'error',
@@ -18,6 +23,11 @@ export const ENTER_VALID_UNI_CODE = {
 export const ENTER_VALID_DNI_CODE = {
   severity: 'error',
   message: 'Por favor, ingrese un documento válido. Ej: 87654321',
+};
+
+export const PASSWORD_SENT_TO_EMAIL_ERROR = {
+  severity: 'error',
+  message: 'Ocurrió un error al intentar enviar la contraseña',
 };
 
 // helpers