+++ src/tools/index.js
@@ -1,4 +1,5 @@
 export * from './request';
 export * from './cookies';
 export * from './object';
+export * from './security';
 export * from './alertSnackbar';